+++ libs/homekit-rsf/src/debug.c
@@ -3,10 +3,10 @@
 #include "debug.h"
 
 
-char *binary_to_string(const byte *data, size_t size) {
+char *binary_to_string(const byte *data, unsigned int size) {
     unsigned int i;
 
-    size_t buffer_size = 1; // 1 char for eos
+    unsigned int buffer_size = 1; // 1 char for eos
     for (i = 0; i < size; i++)
         buffer_size += (data[i] == '\\') ? 2 : ((data[i] >= 32 && data[i] < 128) ? 1 : 4);
 
@@ -18,7 +18,7 @@ char *binary_to_string(const byte *data, size_t size) {
             buffer[pos++] = '\\';
             buffer[pos++] = '\\';
         } else if (data[i] < 32 || data[i] >= 128) {
-            size_t printed = snprintf(&buffer[pos], buffer_size - pos, "\\x%02X", data[i]);
+            unsigned int printed = snprintf(&buffer[pos], buffer_size - pos, "\\x%02X", data[i]);
             if (printed > 0) {
                 pos += printed;
             }
@@ -32,7 +32,7 @@ char *binary_to_string(const byte *data, size_t size) {
 }
 
 
-void print_binary(const char *prompt, const byte *data, size_t size) {
+void print_binary(const char *prompt, const byte *data, unsigned int size) {
 #ifdef HOMEKIT_DEBUG
     char *buffer = binary_to_string(data, size);
     printf("%s (%d bytes): \"%s\"\n", prompt, (int)size, buffer);
